Make the Stream property virtual
This way, sub classes can add their own logic to create the save file.

The common scenario is to compute a checksum.

For Crash Bandicoot 2, instead of computing the checksum every time a
Set*() function is called, we compute it only when necessary, that is,
when the consumer gets the Stream property to create or edit the save file.